Revert "Wip/marzavala/waiter layout v2"
