+++ burguerqueen/src/App.js
@@ -1,22 +1,20 @@
 // import queen from './queen.png';
 import './App.css';
-import { useState } from 'react';
+import {useState} from 'react'
 import {
   RouterProvider,
   Route,
-} from 'react-router-dom';
+} from "react-router-dom";
 import { routerNoAuth } from './components/noAuth/routerNoAuth';
-import { routerAuth } from './components/auth/routerAuth';
+import {routerAuth} from './components/auth/routerAuth'
+
 
 function App() {
-  const [user, setUser] = useState('pao');
-  // use efect setUser
+  const [user, setUser]= useState("pao")
+  //use efect setUser
   return (
-    <div>
-      {' '}
-      {user ? <RouterProvider router={routerAuth} /> : <RouterProvider router={routerNoAuth} />}
-    </div>
-  /*
+   <div> {user?<RouterProvider router={routerAuth} />:<RouterProvider router={routerNoAuth} />}</div>
+    /*
     <div className="App">
       <header className="App-header">
         <img src={queen} className="App-logo" alt="logo" />
@@ -24,10 +22,10 @@ function App() {
           Burger Queen WorkSpace
         </p>
         <p>
-          Coming soon ...
+          Coming soon ... 
         </p>
       </header>
-    </div> */
+    </div>*/
   );
 }
