+++ burguerqueen/src/components/auth/routerAuth.js
@@ -1,17 +1,18 @@
+
 import {
-  createBrowserRouter,
-} from 'react-router-dom';
-// import Welcome from "./Welcome";
-import Menu from './waiter/TakeOrder';
-import Admin from './admin/Admin';
+    createBrowserRouter,
+  } from "react-router-dom";
+//import Welcome from "./Welcome";
+import Menu from "./waiter/TakeOrder";
+import Admin from './admin/Admin'
 
 export const routerAuth = createBrowserRouter([
-  {
-    path: '/',
-    element: <Menu />,
-  },
-  {
-    path: '/admin',
-    element: <Admin />,
-  },
-]);
+    {
+      path: "/",
+      element: <Menu />
+    },
+    {
+        path: "/admin",
+        element: <Admin />
+      },
+  ]);
