--- burguerqueen/src/components/auth/routerAuth.jsx
@@ -1,17 +1,0 @@
-import {
-  createBrowserRouter,
-} from 'react-router-dom';
-// import Welcome from "./Welcome";
-import Menu from './waiter/TakeOrder';
-import Admin from './admin/Admin';
-
-export const routerAuth = createBrowserRouter([
-  {
-    path: '/',
-    element: <Menu />,
-  },
-  {
-    path: '/admin',
-    element: <Admin />,
-  },
-]);
