--- burguerqueen/src/components/auth/waiter/OrderDetails.js
@@ -1,15 +1,0 @@
-function OrderDetails(props){
-    return(
-        <table>
-            <tr>
-            <td>{props.qty}</td>
-            <td>{props.name}</td>
-            <td>{props.price}</td>
-            <img src={require('../../../images/delete.png')} />
-            </tr>
-        </table>
-    )
-}
-
-export default OrderDetails;
-
